FEAT : 카테고리 삭제 3-tier 연동
1. 로컬테스트용으로 작성된 카테고리 삭제 3-tier환경으로 연동
2. 문제 삭제, 카테고리 삭제여부 재확인 confirm alert 추가
3. 삭제 버튼 디자인 변경

+++ frontend/src/views/mypage/Mypage.vue
@@ -118,6 +118,11 @@ export default {
           const userId = localStorage.getItem('user_id'); // 로컬 스토리지에서 사용자 ID 가져오기
           const index = this.add_questions.findIndex(add_questions => add_questions.id === questionId);
           if (index !== -1){
+              const confirmDelete = confirm('정말로 이 문제를 삭제하시겠습니까?');
+              if (!confirmDelete) {
+                // 사용자가 '취소'를 클릭한 경우, 함수를 종료합니다.
+                return;
+              }
               axios.delete(`${process.env.VUE_APP_BACKEND_API}/question/delete?_id=${questionId}&user_id=${userId}`)
                   .then(response => {
                       if (response.status === 200){
